first step of shared module
